requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=1.4.0
numpy>=1.22.0
scikit-learn>=1.1.0
xgboost>=1.7.0
shap>=0.41.0
python-dotenv>=0.21.0
requests>=2.28.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
certifi>=2022.6.15
pydantic>=1.10.0
streamlit>=1.18.0
